+++ src/plugins/rosterview/index.js
@@ -24,7 +24,6 @@ converse.plugins.add('converse-rosterview', {
     initialize () {
         api.settings.extend({
             'autocomplete_add_contact': true,
-            'allow_chat_pending_contacts': true,
             'allow_contact_removal': true,
             'hide_offline_users': false,
             'roster_groups': true,